Extract stable isotope data from a data frame

This function generates a list of isotopeData class objects parsing a data frame of stable isotope values analysed for one or more consumers and one or two baselines. The data frame can be organized in one or more communities (or sampling sites, samples in time, multiple studies, etc.).

Usage

extractIsotopeData(
  df = NULL,
  b1 = "Baseline 1",
  b2 = NULL,
  baselineColumn = "FG",
  consumersColumn = "Spp",
  groupsColumn = NULL,
  deltaC = NULL,
  deltaN = NULL,
  d13C = "d13C",
  d15N = "d15N",
  seed = 3,
  ...
)

Arguments

df

data frame containing raw isotope data, with one or more grouping variables.

b1

string or vector with the text for baseline 1.

b2

string or vector with the text for baseline 2.

baselineColumn

string of the column where baselines are grouped.

consumersColumn

string of the column where consumers/species are grouped.

groupsColumn

string of the column where groups/communities are grouped.

deltaC

vector of values with trophic discrimination factor for carbon. If NULL it will use Post's assumptions (56 values with 3.4 mean +- 0.98 sd).

deltaN

vector of values with trophic discrimination factor for nitrogen. If NULL it will use Post's assumptions (107 values with 0.39 mean +- 1.3 sd).

d13C

string of the column that has d13C isotope values.

d15N

string of the column that has d15N isotope values.

seed

integer to get reproducible results. By default, seed = 3.

...

Additional arguments passed to this funcion.

Value

a list with isotopeData class objects

Examples

data("Bilagay")
head(Bilagay)
#>   Study Location                       Spp         FG  d13C d15N NS
#> 1   MEL      CHI Echinolittorina peruviana Benthic_BL -10.1 16.3  4
#> 2   MEL      CHI Echinolittorina peruviana Benthic_BL -11.7 16.3  4
#> 3   MEL      CHI Echinolittorina peruviana Benthic_BL  -9.8 16.4  4
#> 4   MEL      CHI Echinolittorina peruviana Benthic_BL -11.0 16.6  4
#> 5   MEL      CHI Echinolittorina peruviana Benthic_BL  -7.4 12.5  4
#> 6   MEL      CHI Echinolittorina peruviana Benthic_BL -10.1 14.2  4
isotopeList <- extractIsotopeData(Bilagay, b1 = "Benthic_BL",
b2 = "Pelagic_BL", baselineColumn = "FG", consumersColumn = "Spp",
groupsColumn = "Location", d13C = "d13C", d15N = "d15N")