Extract and load stable isotope data for selected consumers from a data frame

This function extracts only selected consumers/species with their respective baseline(s) and returns an isotopeData class object (or list). It is useful when there are a lot of information in a data frame and you want to calculate trophic position only for selected consumers in one or more communities.

Usage

loadIsotopeData(
  df = NULL,
  consumer = NULL,
  group = NULL,
  b1 = "Baseline 1",
  b2 = NULL,
  baselineColumn = "FG",
  consumersColumn = "FG",
  groupsColumn = NULL,
  d13C = "d13C",
  d15N = "d15N",
  deltaC = NULL,
  deltaN = NULL,
  seed = 666,
  ...
)

Arguments

df

data frame containing raw isotope data with at least one grouping column.

consumer

string or character vector indicating which consumer/species will be extracted.

group

string or character vector indicating which group(s) will be extracted.

b1

string or character vector indicating which baseline(s) will be extracted as baseline 1.

b2

string or character vector indicating which baseline(s) will be extracted as baseline 2.

baselineColumn

string of the column where baselines are grouped.

consumersColumn

string of the column where species/consumer(s) are grouped.

groupsColumn

string of the column where groups/communities are grouped.

d13C

string indicating from which column extract d13C isotope values.

d15N

string indicating from which column extract d15N isotope values.

deltaC

vector of values with trophic discrimination factor for carbon. If NULL it will use Post's assumptions (56 values with 3.4 mean +- 0.98 sd).

deltaN

vector of values with trophic discrimination factor for nitrogen. If NULL it will use Post's assumptions (107 values with 0.39 mean +- 1.3 sd).

seed

numerical value to get reproducible results with trophic discrimination factors (because they are simulated each time this function is called). By default, is 3.

...

Additional arguments passed to this function.

Value

an isotopeData class object if one consumer and one group are selected. A list of isotopeData class objects if more than one consumer or more than one group are selected.

Examples

data("Bilagay")
head(Bilagay)
#>   Study Location                       Spp         FG  d13C d15N NS
#> 1   MEL      CHI Echinolittorina peruviana Benthic_BL -10.1 16.3  4
#> 2   MEL      CHI Echinolittorina peruviana Benthic_BL -11.7 16.3  4
#> 3   MEL      CHI Echinolittorina peruviana Benthic_BL  -9.8 16.4  4
#> 4   MEL      CHI Echinolittorina peruviana Benthic_BL -11.0 16.6  4
#> 5   MEL      CHI Echinolittorina peruviana Benthic_BL  -7.4 12.5  4
#> 6   MEL      CHI Echinolittorina peruviana Benthic_BL -10.1 14.2  4
loadIsotopeData(df = Bilagay, consumer = "Bilagay", consumersColumn = "FG",
group = c("CHI", "COL"), groupsColumn = "Location",
b1 = "Benthic_BL", b2 = "Pelagic_BL", baselineColumn = "FG")
#> $`CHI-Bilagay`
#> $dNb1
#>  [1] 16.3 16.3 16.4 16.6 12.5 14.2 16.2 16.9 15.2 16.3 17.0 17.5 13.5 14.6 15.7
#> [16] 15.7 16.4 16.6 15.9 16.0 16.1 17.5 17.6 17.6 18.2 17.5 18.0 18.1 18.4
#> 
#> $dCb1
#>  [1] -10.1 -11.7  -9.8 -11.0  -7.4 -10.1 -10.9 -11.8  -8.7 -10.0 -12.1 -12.2
#> [13]  -8.5 -10.1  -6.5  -8.0 -14.6 -14.3 -11.4  -6.3  -9.1 -10.4 -10.8 -10.9
#> [25]  -8.0 -11.9 -12.4  -9.9 -12.1
#> 
#> $dNb2
#>  [1] 16.3 16.4 16.4 16.8 15.6 15.9 16.3 16.5 15.7 15.7 16.1 16.5 16.2 16.2 16.8
#> [16] 16.5 16.3 15.6 15.8 16.0 15.7 15.4 15.9 15.9 16.9 16.3
#> 
#> $dCb2
#>  [1] -14.5 -14.5 -14.3 -14.4 -15.1 -15.0 -14.4 -14.5 -14.8 -14.6 -14.7 -14.8
#> [13] -16.3 -15.1 -14.7 -14.9 -14.7 -14.6 -15.6 -14.6 -15.0 -14.8 -15.8 -16.0
#> [25] -15.2 -15.5
#> 
#> $deltaN
#>  [1] 4.145940 5.274408 3.154027 5.286769 1.488014 4.150490 2.302961 2.753676
#>  [9] 1.868005 3.434212 5.395831 1.887701 4.245577 1.932512 3.591850 3.403970
#> [17] 4.239891 3.780466 2.950607 4.175344 2.852487 2.413157 4.607280 3.192133
#> [25] 3.499183 2.145384 2.463623 1.893417 2.520921 2.270017 4.147448 2.897777
#> [33] 4.752501 2.912934 3.676746 3.707337 3.855919 2.143643 3.632909 5.316203
#> [41] 4.121143 3.034473 3.038600 5.123389 3.933071 2.474237 3.194283 3.553644
#> [49] 4.048405 3.807253 3.648585 4.529300 2.844831 2.928922 3.019297 2.737296
#> 
#> $deltaC
#>   [1]  1.41949496  2.98997741  0.03905562  3.00718011 -2.27952397  1.42582789
#>   [7] -1.14536701 -0.51811025 -1.75069243  0.42898757  3.15896090 -1.72328155
#>  [13]  1.55815909 -1.66091906  0.64837193  0.38690096  1.55024711  0.91086716
#>  [19] -0.24404267  1.46041741 -0.38059470 -0.99200838  2.06153935  0.09208766
#>  [25]  0.51940701 -1.36466609 -0.92177489 -1.71532660 -0.84203380 -1.19121452
#>  [31]  1.42159357 -0.31756537  2.26364215 -0.29647084  0.76652039  0.80909467
#>  [37]  1.01587531 -1.36708863  0.70551348  3.04814387  1.38498611 -0.12732688
#>  [43] -0.12158237  2.77980521  1.12324716 -0.90700332  0.09507985  0.59520130
#>  [49]  1.28375595  0.94814663  0.72732998  1.95301552 -0.39124939 -0.27422118
#>  [55] -0.14844602 -0.54090568  0.18727280 -0.47724180  1.37800292  2.07536287
#>  [61] -0.75617803 -0.64709350 -1.16036777  2.89052068  1.32578179 -0.01990255
#>  [67]  0.71240732  0.97929380  2.52873827  0.40393477  0.08727896  0.29992536
#>  [73]  1.35759460  0.72005565 -1.04784671  0.39921068 -0.49738347 -3.18406996
#>  [79] -0.30034219 -0.20756533  2.33647162  1.58678372  0.07234983 -1.11692355
#>  [85]  0.46341966  1.73336500  0.98336403  2.26535549  0.65237642 -1.18935619
#>  [91] -0.12514745 -0.63946274  0.88045553  0.87329342 -0.81485288  0.96135980
#>  [97]  1.47459535 -0.20823681  0.70215697  0.53486693 -0.69855100 -1.78116678
#> [103]  2.95414402  1.49851761 -0.76360248  1.18014627 -0.49012864
#> 
#> $dNc
#>  [1] 21.6 21.7 22.0 21.6 21.8 21.9 22.0 21.3 22.3 21.7 21.9 22.2 21.5 22.4 22.4
#> [16] 22.5 22.3 22.1
#> 
#> $dCc
#>  [1] -13.6 -14.5 -13.8 -14.4 -14.4 -14.4 -14.4 -14.1 -14.5 -14.3 -14.3 -12.7
#> [13] -14.2 -14.2 -15.0 -14.1 -14.4 -11.9
#> 
#> attr(,"class")
#> [1] "isotopeData"
#> attr(,"consumer")
#> [1] "Bilagay"
#> attr(,"baseline1")
#> [1] "Benthic_BL"
#> attr(,"baseline2")
#> [1] "Pelagic_BL"
#> attr(,"group")
#> [1] "CHI"
#> 
#> $`COL-Bilagay`
#> $dNb1
#>  [1] 15.7 16.8 17.0 17.1 17.4 17.5 18.4 18.5 13.7 15.9 16.3 17.1 15.4 16.0 16.2
#> [16] 16.2 16.0 16.3 16.4 17.4 17.9 18.1 18.2 18.3 18.3 18.5 18.6 18.6 17.3 17.1
#> [31] 17.1 17.1 17.6 17.1 17.4 17.4 17.1 17.3 17.2 17.5 17.6 17.4
#> 
#> $dCb1
#>  [1]  -6.9  -8.8  -7.7  -8.5 -11.3 -12.8  -9.8  -9.2  -8.6 -11.2 -12.4 -10.0
#> [13]  -6.7 -10.3 -10.8  -9.3  -9.0  -7.7  -7.2  -7.0 -11.7 -11.9 -12.6 -11.4
#> [25] -14.1 -14.1 -13.6 -14.0 -12.5 -13.0 -12.7 -11.7 -13.3 -11.9 -11.7 -12.1
#> [37] -11.4 -13.7 -14.2 -14.2 -12.4 -13.7
#> 
#> $dNb2
#>  [1] 15.5 15.9 16.1 16.3 16.5 16.5 16.9 17.1 15.6 16.1 16.4 17.1 15.2 15.1 14.9
#> [16] 15.0 15.8 16.8 15.7 16.5 16.5 15.5 16.7 16.6 16.2 16.2 16.4 15.3 16.8 16.8
#> [31] 16.8 16.2
#> 
#> $dCb2
#>  [1] -15.2 -14.5 -14.9 -14.5 -14.0 -15.0 -14.9 -14.0 -15.0 -15.0 -14.6 -14.4
#> [13] -14.3 -13.9 -14.5 -14.4 -13.5 -14.9 -15.9 -15.2 -15.0 -15.3 -16.7 -15.4
#> [25] -15.3 -15.4 -14.8 -16.7 -14.1 -14.3 -14.1 -14.9
#> 
#> $deltaN
#>  [1] 4.145940 5.274408 3.154027 5.286769 1.488014 4.150490 2.302961 2.753676
#>  [9] 1.868005 3.434212 5.395831 1.887701 4.245577 1.932512 3.591850 3.403970
#> [17] 4.239891 3.780466 2.950607 4.175344 2.852487 2.413157 4.607280 3.192133
#> [25] 3.499183 2.145384 2.463623 1.893417 2.520921 2.270017 4.147448 2.897777
#> [33] 4.752501 2.912934 3.676746 3.707337 3.855919 2.143643 3.632909 5.316203
#> [41] 4.121143 3.034473 3.038600 5.123389 3.933071 2.474237 3.194283 3.553644
#> [49] 4.048405 3.807253 3.648585 4.529300 2.844831 2.928922 3.019297 2.737296
#> 
#> $deltaC
#>   [1]  1.41949496  2.98997741  0.03905562  3.00718011 -2.27952397  1.42582789
#>   [7] -1.14536701 -0.51811025 -1.75069243  0.42898757  3.15896090 -1.72328155
#>  [13]  1.55815909 -1.66091906  0.64837193  0.38690096  1.55024711  0.91086716
#>  [19] -0.24404267  1.46041741 -0.38059470 -0.99200838  2.06153935  0.09208766
#>  [25]  0.51940701 -1.36466609 -0.92177489 -1.71532660 -0.84203380 -1.19121452
#>  [31]  1.42159357 -0.31756537  2.26364215 -0.29647084  0.76652039  0.80909467
#>  [37]  1.01587531 -1.36708863  0.70551348  3.04814387  1.38498611 -0.12732688
#>  [43] -0.12158237  2.77980521  1.12324716 -0.90700332  0.09507985  0.59520130
#>  [49]  1.28375595  0.94814663  0.72732998  1.95301552 -0.39124939 -0.27422118
#>  [55] -0.14844602 -0.54090568  0.18727280 -0.47724180  1.37800292  2.07536287
#>  [61] -0.75617803 -0.64709350 -1.16036777  2.89052068  1.32578179 -0.01990255
#>  [67]  0.71240732  0.97929380  2.52873827  0.40393477  0.08727896  0.29992536
#>  [73]  1.35759460  0.72005565 -1.04784671  0.39921068 -0.49738347 -3.18406996
#>  [79] -0.30034219 -0.20756533  2.33647162  1.58678372  0.07234983 -1.11692355
#>  [85]  0.46341966  1.73336500  0.98336403  2.26535549  0.65237642 -1.18935619
#>  [91] -0.12514745 -0.63946274  0.88045553  0.87329342 -0.81485288  0.96135980
#>  [97]  1.47459535 -0.20823681  0.70215697  0.53486693 -0.69855100 -1.78116678
#> [103]  2.95414402  1.49851761 -0.76360248  1.18014627 -0.49012864
#> 
#> $dNc
#>  [1] 21.2 21.6 21.3 21.1 20.8 21.1 21.0 21.4 21.7 21.3 21.5 21.8 21.0 21.2 20.9
#> [16] 21.3 21.6 22.0 20.8 21.3 21.8 21.5 21.0 21.2 21.1 22.1 21.5 21.0 21.1 21.3
#> [31] 21.3 20.5 20.8 21.4 21.1 19.9 20.9 21.2 20.5 21.0 21.0 21.0
#> 
#> $dCc
#>  [1] -15.5 -15.6 -14.8 -15.1 -15.0 -15.2 -15.3 -15.0 -15.0 -14.9 -14.2 -14.8
#> [13] -15.7 -15.2 -15.2 -15.4 -15.2 -15.0 -15.2 -15.6 -15.0 -15.3 -15.0 -15.1
#> [25] -14.5 -15.1 -15.8 -13.1 -13.1 -13.3 -12.8 -13.2 -14.5 -13.5 -14.0 -13.2
#> [37] -15.6 -13.1 -14.1 -13.2 -13.8 -12.4
#> 
#> attr(,"class")
#> [1] "isotopeData"
#> attr(,"consumer")
#> [1] "Bilagay"
#> attr(,"baseline1")
#> [1] "Benthic_BL"
#> attr(,"baseline2")
#> [1] "Pelagic_BL"
#> attr(,"group")
#> [1] "COL"
#>