Function that summarises a data frame containing stable isotope values (d13C and d15N) grouping by Species and FG columns

A wrapper of plyr:ddply to summarise a data frame.

Usage

summariseIsotopeData(
  df = NULL,
  grouping = c("Species", "FG"),
  printSummary = FALSE,
  ...
)

Arguments

df

a data frame that contains the isotope values. It needs to have the following columns: d13C, d15N, Species and FG. Species stands for the scientific name (or common name), and FG stands for the functional group for each species. If the data frame does not have Species and FG columns, it will raise an error. If the columns change their names, they need to be stated as well in the grouping variable.

grouping

a vector with the name of the columns (variables) that will be used to summarize, and plot the data frame.

printSummary

logical value indicating whether the summary is printed.

...

optional arguments that are passed to the function for later use.

Value

a data frame with the summary of the data frame.

Examples

data('Bilagay')
subset_CHI <- Bilagay[Bilagay[,'Location'] %in% 'CHI',]
summariseIsotopeData(subset_CHI, grouping = c('Spp', 'FG'))
#>                         Spp         FG nC      meanC       sdC       seC nN
#> 1                   Bilagay    Bilagay 18 -14.066667 0.7194769 0.1695823 18
#> 2 Echinolittorina peruviana Benthic_BL 12 -10.483333 1.4427457 0.4164848 12
#> 3    Perumytilus purpuratus Pelagic_BL 26 -14.938462 0.5192895 0.1018410 26
#> 4          Scurria viridula Benthic_BL  9  -9.866667 3.0508196 1.0169399  9
#> 5                    Tegula Benthic_BL  8 -10.800000 1.4282857 0.5049752  8
#>      meanN       sdN        seN
#> 1 21.95556 0.3484794 0.08213739
#> 2 15.95000 1.3813695 0.39876704
#> 3 16.14231 0.4051401 0.07945450
#> 4 15.61111 0.9726824 0.32422748
#> 5 17.86250 0.3543102 0.12526757